:recycle: [REFACTOR] header text 컴포넌트 분리
