+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import './style/globals.css';
 import StartPage from './components/StartPage';
