+++ src/components/Header/Header.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import { AiOutlineMenu } from 'react-icons/ai';
 import { BsPersonCircle } from 'react-icons/bs';
-import LoginModal from './Login/LoginModal';
+import LoginModal from '../Login/LoginModal';
+import HeaderText from './HeaderText';
 
 export default function Header() {
   const [isOpenModal, setIsOpenModal] = useState(false)
@@ -16,11 +17,11 @@ export default function Header() {
         </a>
       </div>
       <div className='flex items-center justify-start w-4/5 text-base font-bold gap-x-12 grow'>
-        <a href='#'>변수명 추천</a>
-        <a href='#'>Code 언어 변경</a>
-        <a href='#'>Code 리팩토링</a>
-        <a href='#'>Code 주석 처리</a>
-        <a href='#'>컨텐츠 추천</a>
+        <HeaderText text={'변수명 추천'} />
+        <HeaderText text={'Code 언어 변경'} />
+        <HeaderText text={'Code 리팩토링'} />
+        <HeaderText text={'Code 주석처리'} />
+        <HeaderText text={'컨텐츠 추천'} />
       </div>
       <div className='flex items-center justify-end w-[10%] text-orange-700' onClick={()=>setIsOpenModal(true)}>
         <a href="#" className="flex items-center justify-center w-4/5 gap-2 mr-4 border border-orange-500 rounded-lg h-3/5">
