+++ src/components/Header/HeaderText.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function HeaderText({ text }) {
+  return (
+    <a href='#' className="menu-text-hover">{text}</a>
+  )
+}
