+++ src/style/globals.css
@@ -6,6 +6,10 @@
   @apply flex justify-center items-center
 }
 
+.menu-text-hover {
+  @apply hover:text-[#3B82F6]
+}
+
 .co-re-color {
   @apply text-[#3B82F6]
 }
